Taking into account the 'Dicom Sequences' leads up to introduce new concepts (and new classes) : a 'gdcmDocument' is composed of a set of Doc Entries, that are   - elementary gdcmDocEntries (former gdcmHeaderEntries)   - Sequence Doc Entries (gdcmSeqEntries)     a Sequence is composed of Items.     Each item is a set of Doc Entries (this is recursive)
The non uniqueness of the 'Dicom tag' is due to this recursivity
(never taken into account)
Our unability to add a new 'entry' in the header
    at the proper location (neither in the H-Table (map or multimap), nor in
    the Chained List is also due to this recursivity.

Don't try, right now, to use/modify/compile these new sources : nothing is
finished.
We just commit all the stuff, 'as is', in order not to loose it.